Extract common steps and helper methods into a separate file
